pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.5",
    "fastapi>=0.116.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "python-socketio>=5.13.0",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]
